Make many tests in gedbrowser-mongo-dao be integration tests
