+++ gedbrowser-mongo-dao/src/test/java/org/schoellerfamily/gedbrowser/persistence/mongo/loader/test/GedDocumentFileLoaderIT.java
@@ -25,7 +25,7 @@ import org.springframework.test.context.junit.jupiter.SpringExtension;
  */
 @ExtendWith(SpringExtension.class)
 @ContextConfiguration(classes = { MongoTestConfiguration.class })
-public class GedDocumentFileLoaderTest {
+public class GedDocumentFileLoaderIT {
     /** */
     @Autowired
     private transient GedDocumentFileLoader loader;
